+++ src/components/index.js
@@ -1,7 +1,7 @@
-export { default as About } from './About';
-export { default as Projects } from './Projects';
+export { default as About } from './About/About';
+export { default as Projects } from './Projects/Projects';
 export { default as NavBar } from './NavBar';
-export { default as Experience } from './Experience';
-export { default as Education } from './Education';
-export { default as Awards } from './Awards';
-export { default as Skills } from './Skills';
+export { default as Experience } from './Experience/Experience';
+export { default as Education } from './Education/Education';
+export { default as Awards } from './Awards/Awards';
+export { default as Skills } from './Skills/Skills';
